+++ src/arena_internal.h
@@ -3,8 +3,8 @@
 
 struct arena {
     unsigned char* buffer;
-    size_t size;
-    size_t next_offset;
+    int32_t size;
+    int32_t next_offset;
 };
 
 #endif //ARENA_MEMTABLE_ARENA_INTERNAL_H
